+++ lib/packages/wikibase/data-model-serialization/RELEASE-NOTES.md
@@ -1,5 +1,10 @@
 # Wikibase DataModel Serialization release notes
 
+## 2.1.0 (dev)
+
+* Added `newItemSerializer` and `newPropertySerializer` to `SerializerFactory`
+* Added `newItemDeserializer` and `newPropertyDeserializer` to `DeserializerFactory`
+
 ## 2.0.0 (2015-08-30)
 
 * Dropped dependency on Wikibase DataModel Services